# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GetPraying API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Now following
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  following:
                    type: boolean
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  following:
                    type: boolean
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by a specific user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /users/{username}/saved-posts:
    get:
      operationId: getUserSavedPosts
      tags: [users]
      summary: Get a user's saved feed posts (approved only, most recently saved first)
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: Saved posts visible on the user's public Saved tab
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPostsList"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # POSTS / FEED
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        prayersShared:
          type: integer
        prayedFor:
          type: integer
        savedScrolls:
          type: integer
        followerCount:
          type: integer
        followingCount:
          type: integer
        isFollowing:
          type: boolean
          description: Present when the viewer is authenticated and not viewing their own profile
        scheduledNotificationsEnabled:
          type: boolean
          nullable: true
          description: Present only when the API includes viewer-specific settings for this profile
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - prayersShared
        - prayedFor
        - savedScrolls
        - followerCount
        - followingCount
        - createdAt
    Post:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        mediaUrl:
          type: string
          nullable: true
        mediaType:
          type: string
          nullable: true
          enum: [image, video, audio]
        category:
          type: string
          nullable: true
        categories:
          type: array
          items:
            type: string
          description: All tag slugs for this post (primary is usually first)
        isAnonymous:
          type: boolean
        status:
          type: string
          enum: [pending, approved, declined]
        flagReason:
          type: string
          nullable: true
        moderationReason:
          type: string
          nullable: true
          description: Shown to the author when a moderator declines the post
        prayCount:
          type: integer
        commentCount:
          type: integer
        saveCount:
          type: integer
        hasPrayed:
          type: boolean
        hasCommented:
          type: boolean
        isSaved:
          type: boolean
        authorId:
          type: integer
          nullable: true
        authorUsername:
          type: string
          nullable: true
        authorDisplayName:
          type: string
          nullable: true
        authorAvatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        boostedAt:
          type: string
          format: date-time
          nullable: true
          description: When a premium member last boosted this post (feed priority)
        boostedByUserId:
          type: integer
          nullable: true
          description: User id of the member whose boost is currently active (for unboost toggle)
      required:
        - id
        - content
        - isAnonymous
        - status
        - prayCount
        - commentCount
        - saveCount
        - hasPrayed
        - hasCommented
        - isSaved
        - createdAt
...
    SavedPostsList:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required:
        - posts
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        nextCursor:
          type: string
          nullable: true
          description: Opaque feed pagination cursor — pass as GET /posts?cursor=
        total:
          type: integer
      required:
        - posts
        - total
...